Organize i18n folder structure
