--- src/i18n/index.de.json
@@ -1,11 +1,0 @@
-{
-  "greeting": "Hallo Welt!",
-
-  "navbar": {
-    "tricks": "Tricks",
-    "combos": "Kombos",
-    "glossary": "Glossar",
-    "generator": "Generator",
-    "more": "Mehr"
-  }
-}
