--- src/i18n/index.en.json
@@ -1,11 +1,0 @@
-{
-  "greeting": "Hello World!",
-
-  "navbar": {
-    "tricks": "Tricks",
-    "combos": "Combos",
-    "glossary": "Glossary",
-    "generator": "Generator",
-    "more": "More"
-  }
-}
