--- src/i18n/index.es.json
@@ -1,11 +1,0 @@
-{
-  "greeting": "Hola mundo!",
-
-  "navbar": {
-    "tricks": "Trucos",
-    "combos": "Combos",
-    "glossary": "Glosario",
-    "generator": "Generador",
-    "more": "Más"
-  }
-}
