+++ src/i18n/index/index.de.json
@@ -1,0 +1,3 @@
+{
+  "greeting": "Hallo Welt!"
+}
