+++ src/i18n/index/index.en.json
@@ -1,0 +1,3 @@
+{
+  "greeting": "Hello World!"
+}
