+++ src/i18n/index/index.es.json
@@ -1,0 +1,3 @@
+{
+  "greeting": "Hola mundo!"
+}
