+++ src/i18n/index/index.fr.json
