+++ src/i18n/navbar/navbar.de.json
@@ -1,0 +1,7 @@
+{
+  "tricks": "Tricks",
+  "combos": "Kombos",
+  "glossary": "Glossar",
+  "generator": "Generator",
+  "more": "Mehr"
+}
