+++ src/i18n/navbar/navbar.en.json
@@ -1,0 +1,7 @@
+{
+  "tricks": "Tricks",
+  "combos": "Combos",
+  "glossary": "Glossary",
+  "generator": "Generator",
+  "more": "More"
+}
