+++ src/i18n/navbar/navbar.es.json
@@ -1,0 +1,7 @@
+{
+  "tricks": "Trucos",
+  "combos": "Combos",
+  "glossary": "Glosario",
+  "generator": "Generador",
+  "more": "Más"
+}
