+++ src/i18n/navbar/navbar.fr.json
@@ -1,0 +1,2 @@
+{
+}
