+++ src/i18n/navbar/navbar.ts
@@ -1,0 +1,8 @@
+import de from "./navbar.de.json"
+import en from "./navbar.en.json"
+import fr from "./navbar.fr.json"
+import es from "./navbar.es.json"
+
+export default ({
+  de, en, fr, es
+})
